generator_block/requirements.txt
kafka-python
websocket
websocket-client
datetime